requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
python-dotenv
pydantic-settings
Pillow
langchain
langchain-groq
langchain-google-genai
chromadb
sentence-transformers
torch
numpy
